// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
export interface ProjectDocumentDataBadgesItem {
  /**
   * Label field in *Project → Badges*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: project.badges[].label
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  label: prismic.KeyTextField;

  /**
   * Color field in *Project → Badges*
   *
   * - **Field Type**: Select
   * - **Placeholder**: *None*
   * - **API ID Path**: project.badges[].color
   * - **Documentation**: https://prismic.io/docs/field#select
   */
  color: prismic.SelectField<
    | "whiteAlpha"
    | "blackAlpha"
    | "gray"
    | "red"
    | "orange"
    | "yellow"
    | "green"
    | "teal"
    | "blue"
    | "cyan"
    | "purple"
    | "pink"
  >;
}

/**
 * Item in *Project → Buttons*
 */
export interface ProjectDocumentDataButtonsItem {
  /**
   * Link field in *Project → Buttons*
   *
   * - **Field Type**: Link
   * - **Placeholder**: *None*
   * - **API ID Path**: project.buttons[].link
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  link: prismic.LinkField;

  /**
   * Label field in *Project → Buttons*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: project.buttons[].label
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  label: prismic.KeyTextField;
}

/**
 * Content for Project documents
 */
interface ProjectDocumentData {
  /**
   * SEO Name field in *Project*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: project.seo_name
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  seo_name: prismic.KeyTextField;

  /**
   * SEO Description field in *Project*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: project.seo_description
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  seo_description: prismic.KeyTextField;

  /**
   * Image field in *Project*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: project.image
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  image: prismic.ImageField<never>;

  /**
   * Badges field in *Project*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: project.badges[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  badges: prismic.GroupField<Simplify<ProjectDocumentDataBadgesItem>>;

  /**
   * Buttons field in *Project*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: project.buttons[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  buttons: prismic.GroupField<Simplify<ProjectDocumentDataButtonsItem>>;

  /**
   * Fixed field in *Project*
   *
   * - **Field Type**: Boolean
   * - **Placeholder**: *None*
   * - **Default Value**: false
   * - **API ID Path**: project.fixed
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#boolean
   */
  fixed: prismic.BooleanField;
}

/**
 * Project document from Prismic
 *
 * - **API ID**: `project`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type ProjectDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<
    Simplify<ProjectDocumentData>,
    "project",
    Lang
  >;